Monte Carlo works!  But much slower on before-64 than mid-64.  Surprising.
It produces the same answer for mid-elite8, which isn't a great test since there
are only 5 games left.

Also, it's much slower when run on before-roundof64 than mid-roundof64, even
though it should be doing the same number of pairs (1 million).  So either it's
not doing the same number of pairs, or the bottleneck has moved to the prep for
the pairs, or pairs are taking longer for some reason (memory allocation?)  So
next step is to profile.